implemented the liquor listings, ratings, reviews and prices
Clicking on a Liquor category now shows you all liquors in that category, the average price and the average rating, still waiting on images (only rum works right now with a placeholder image). Clicking on the liquor name shows an image, rating, and one randomly chosen review with the option to see all reviews.

Also tweaked Sam's HTML and CSS to fit with the Design Document and database restrictions.